# apps/api/openapi-v2.yml
openapi: 3.0.3
info:
  title: Edge-Cloud Orchestrator API
  description: Production API documentation for the Edge-Cloud Orchestrator Control Plane
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v2/workflows/:
    get:
      summary: List workflows
      tags:
        - workflows
      responses:
        "200":
          description: Default Response
    post:
      summary: Create a new workflow
      tags:
        - workflows
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                version:
                  type: string
                  pattern: ^\d+\.\d+\.\d+$
                nodes:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                        minLength: 1
                        maxLength: 100
                        pattern: ^[a-zA-Z0-9\-_]+$
                      name:
                        type: string
                        minLength: 1
                        maxLength: 100
                      type:
                        type: string
                        enum:
                          - task
                          - decision
                          - parallel
                          - wait
                          - subworkflow
                      config:
                        type: object
                        additionalProperties: {}
                        propertyNames:
                          maxLength: 100
                      inputs:
                        type: array
                        items:
                          type: string
                      outputs:
                        type: array
                        items:
                          type: string
                    required:
                      - id
                      - name
                      - type
                      - config
                      - inputs
                      - outputs
                    additionalProperties: false
                  minItems: 1
                edges:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      from:
                        type: string
                      to:
                        type: string
                      condition:
                        type: string
                    required:
                      - id
                      - from
                      - to
                    additionalProperties: false
                variables:
                  type: object
                  additionalProperties: {}
                timeout:
                  type: integer
                  minimum: 1000
                  maximum: 86400000
                  default: 60000
                retryPolicy:
                  type: object
                  properties:
                    maxRetries:
                      type: integer
                      minimum: 0
                      maximum: 10
                      default: 3
                    initialDelay:
                      type: integer
                      minimum: 100
                      default: 1000
                    maxDelay:
                      type: integer
                      minimum: 1000
                      default: 60000
                    multiplier:
                      type: number
                      minimum: 1
                      default: 2
                  additionalProperties: false
              required:
                - name
                - version
                - nodes
                - edges
              additionalProperties: false
        required: true
      responses:
        "200":
          description: Default Response
  /v2/workflows/executions/{executionId}:
    get:
      summary: Get workflow execution status
      tags:
        - workflows
      parameters:
        - schema:
            type: string
          in: path
          name: executionId
          required: true
      responses:
        "200":
          description: Default Response
  /v2/workflows/{id}:
    get:
      summary: Get workflow by ID
      tags:
        - workflows
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /v2/workflows/{id}/execute:
    post:
      summary: Execute a workflow
      tags:
        - workflows
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
                  additionalProperties: {}
              additionalProperties: false
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response